css/css-fonts/parsing/font-width-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 4: parsing font-width with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-4/#font-width-prop">
<meta name="assert" content="font-width supports the full grammar 'normal | <percentage [0,∞]> | ultra-condensed | extra-condensed | condensed | semi-condensed | semi-expanded | expanded | extra-expanded | ultra-expanded'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value('font-width', 'normal');
test_valid_value('font-width', 'ultra-condensed');
test_valid_value('font-width', 'extra-condensed');
test_valid_value('font-width', 'condensed');
test_valid_value('font-width', 'semi-condensed');
test_valid_value('font-width', 'semi-expanded');
test_valid_value('font-width', 'expanded');
test_valid_value('font-width', 'extra-expanded');
test_valid_value('font-width', 'ultra-expanded');

test_valid_value('font-width', '234.5%');
test_valid_value('font-width', 'calc(100%)');
test_valid_value('font-width', 'calc(0%)');
test_valid_value('font-width', 'calc(-100%)');
test_valid_value('font-width', 'calc(100% + 100%)', 'calc(200%)');
test_valid_value('font-width', 'calc(100% + (sign(20cqw - 10px) * 5%))', 'calc(100% + (5% * sign(20cqw - 10px)))');
</script>
</body>
</html>
